# docs/static/assets/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/diagram/{diagramId}": {
      "get": {
        "tags": [
          "diagram-controller"
        ],
        "operationId": "getDiagram",
        "parameters": [
          {
            "name": "diagramId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DiagramTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "diagram-controller"
        ],
        "operationId": "updateDiagram",
        "parameters": [
          {
            "name": "diagramId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiagramUpdateTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DiagramTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "diagram-controller"
        ],
        "summary": "Delete one Diagram and all of its versions",
        "operationId": "deleteDiagram",
        "parameters": [
          {
            "name": "diagramId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/version/{diagramId}": {
      "post": {
        "tags": [
          "diagram-version-controller"
        ],
        "operationId": "createOrUpdateVersion",
        "parameters": [
          {
            "name": "diagramId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiagramVersionUploadTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/diagram/{repositoryId}": {
      "post": {
        "tags": [
          "diagram-controller"
        ],
        "operationId": "createDiagram",
        "parameters": [
          {
            "name": "repositoryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewDiagramTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DiagramTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/version/{diagramId}/{versionId}": {
      "get": {
        "tags": [
          "diagram-version-controller"
        ],
        "operationId": "getSingleVersion",
        "parameters": [
          {
            "name": "diagramId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "versionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DiagramVersionTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/version/{diagramId}/{versionId}/download": {
      "get": {
        "tags": [
          "diagram-version-controller"
        ],
        "operationId": "downloadVersion",
        "parameters": [
          {
            "name": "diagramId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "versionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/version/{diagramId}/latest": {
      "get": {
        "tags": [
          "diagram-version-controller"
        ],
        "summary": "Return the latest version of the requested diagram",
        "operationId": "getLatestVersion",
        "parameters": [
          {
            "name": "diagramId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DiagramVersionTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/version/{diagramId}/all": {
      "get": {
        "tags": [
          "diagram-version-controller"
        ],
        "operationId": "getAllVersions",
        "parameters": [
          {
            "name": "diagramId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DiagramVersionTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DiagramUpdateTO": {
        "required": [
          "description",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "fileType": {
            "type": "string"
          },
          "svgPreview": {
            "type": "string"
          }
        }
      },
      "DiagramTO": {
        "required": [
          "description",
          "name",
          "repositoryId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "repositoryId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "updatedDate": {
            "type": "string",
            "format": "date-time"
          },
          "svgPreview": {
            "type": "string"
          },
          "fileType": {
            "type": "string"
          }
        }
      },
...
      "DiagramVersionUploadTO": {
        "type": "object",
        "properties": {
          "versionComment": {
            "type": "string"
          },
          "xml": {
            "type": "string"
          },
          "saveType": {
            "type": "string",
            "enum": [
              "MILESTONE",
              "AUTOSAVE"
            ]
          }
        }
      },
      "NewDiagramTO": {
        "required": [
          "description",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "fileType": {
            "type": "string"
          },
          "svgPreview": {
            "type": "string"
          }
        }
      },
...
      "DiagramVersionTO": {
        "required": [
          "diagramId",
          "repositoryId",
          "saveType",
          "updatedDate",
          "xml"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "comment": {
            "type": "string"
          },
          "milestone": {
            "type": "integer",
            "format": "int32"
          },
          "xml": {
            "type": "string"
          },
          "saveType": {
            "type": "string",
            "enum": [
              "MILESTONE",
              "AUTOSAVE"
            ]
          },
          "updatedDate": {
            "type": "string",
            "format": "date-time"
          },
          "diagramId": {
            "type": "string"
          },
          "repositoryId": {
            "type": "string"
          }
        }
      },
...
    }
  }
}